The my-computer-is-out-of-disk-space Update
Added loads of stuff

+++ src/dev/zt/UpliftedVFFV/inventory/equipables/AtomicPunchThrusters.java
@@ -27,7 +27,7 @@ public class AtomicPunchThrusters extends Item{
 	
 	public status[] getEnchantment(Schmuck s) {
 		enchantment[0] = new ExplosiveCrit((int)(s.getBuffedPow()*(1+s.getEquipPow())),s.itemDummy, 75);
-		enchantment[1] = new BonusStatBuff(16,(int)(.15*(1+s.getEquipPow())),s.itemDummy,75);
+		enchantment[1] = new BonusStatBuff(16,.15*(1+s.getEquipPow()),s.itemDummy,75);
 		return enchantment;
 	}
 }
